requirements.txt
streamlit
google-generativeai
fpdf
pdfkit
